Extract the stack cleaner-upper, for reuse.
