+++ local-modules/util-common/index.js
@@ -5,6 +5,7 @@
 import ColorSelector from './ColorSelector';
 import ColorUtil from './ColorUtil';
 import DeferredLoader from './DeferredLoader';
+import ErrorUtil from './ErrorUtil';
 import FrozenBuffer from './FrozenBuffer';
 import IterableUtil from './IterableUtil';
 import JsonUtil from './JsonUtil';
@@ -18,6 +19,7 @@ export {
   ColorSelector,
   ColorUtil,
   DeferredLoader,
+  ErrorUtil,
   FrozenBuffer,
   IterableUtil,
   JsonUtil,
